# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EP1 Exam",
    "description": "documentation for Noroff EP - e-commerce"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets list of all categories",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Creates a new category",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postcategory"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Alters existing category of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/altercategory"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Deletes category of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/categories/{categoryid}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets list of category of id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "categoryid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "Gets cart of user",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "Adds product to users shopping cart",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/addtocart"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Cart"
        ],
        "description": "Alters a product in cart/decrement",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/altercart"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "Deletes item from cart",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/deletecart"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cart/checkout/now": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "Checks out users cart and creates order.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "delete": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "id"
      ]
    },
    "addtocart": {
      "type": "object",
      "properties": {
        "productid": {
          "type": "number",
          "example": 1
        },
        "quantity": {
          "type": "number",
          "example": 2
        },
        "unitPrice": {
          "type": "number",
          "example": 49.99
        }
      },
      "required": [
        "productid",
        "quantity",
        "unitPrice"
      ]
    },
    "altercart": {
      "type": "object",
      "properties": {
        "productid": {
          "type": "number",
          "example": 1
        },
        "quantity": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productid"
      ]
    },
    "deletecart": {
      "type": "object",
      "properties": {
        "productid": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productid"
      ]
    },
    "postcategory": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "Hockey jerseys"
        }
      },
      "required": [
        "category"
      ]
    },
    "altercategory": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "Hockey jerseys"
        },
        "id": {
          "type": "number",
          "example": 69
        }
      },
      "required": [
        "category",
        "id"
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": {
    "Admin": [
      "adminJWTtoken"
    ]
  }
}